Intro Dictionary based TensileMergeLibrary
	ps1. rebase with rocm repo and fix merge conflicts
	ps2. undo changes to merge.py (being outdated)
	ps3. fix CI issue on LibraryIO.py.
